Added ERD from QDBD

+++ ETL_final/crowdfunding_db_DDL.sql
@@ -21,7 +21,7 @@ CREATE TABLE campaign(
 	company_name VARCHAR(50) NOT NULL,
 	description VARCHAR(100) NOT NULL,
 	goal FLOAT NOT NULL,
-	pledged INT NOT NULL, 
+	pledged FLOAT NOT NULL, 
 	outcome VARCHAR(50) NOT NULL,
 	backers_count INT NOT NULL,
 	country VARCHAR(5) NOT NULL,
